+++ app/about/Section6/Section6.css
