+++ app/about/Slider/Slider.css
